testing/webintel/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
groq>=1.0.0
tavily-python>=0.3.0
supabase==2.10.0
httpx>=0.27.0
aiohttp>=3.9.0
